requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==22.1.0
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.0
Deprecated==1.2.13
discord==1.7.3
discord.py==1.7.3
idna==3.3
lxRbckl==0.92
multidict==6.0.2
pycparser==2.21
PyGithub==1.55
PyJWT==2.4.0
PyNaCl==1.5.0
requests==2.28.1
typing_extensions==4.3.0
urllib3==1.26.11
wrapt==1.14.1
yarl==1.8.1
